Moved most of quiz stuff into `quiz-utils.ts`
